move NetworkRequestEvent into a helper
